Routing by eigen value method

Functions to setup and then evaluate the eigen vector method of Dummit, 2016

The following notes are taken verbatim from the original implimentation of Peter. They do not appear to match the equations and may be irrelevent but added for completeness... ## Qin = exS ## Linear storage-discharge relationship ## Qout = a * vof * ex ## dExS/dt = Qin - Qbf ## = Wv ## ExS is total surface excess storage ## where Wv = complementary overland routing matrix

## Solution for surface routing by Eigenvalue method (see Dummit, 2016; ) ## general soln at time t is ## sigma(exp(lambda_i * t*)ci * ui) ## where ui are eigenvectors, li eigenvalues and ci constants TBD by ## boundary conditions

eigen_routing_setup(K)

eigen_routing_step(v, eig, dt)

Arguments

v
vector of volumes to be redistributed

eig
an object containing summaries of eigen matrix and vectors

dt
time step to be taken

W
matrix of weights for distribution

vof
UNKNOWN

Value

Either an eig object (eigen_setup) or a revised set of volumes after time step