Functions to resample an xts time series

Takes an xts time series object and resamples then to a new time step.

resample_xts(obs, dt, is.rate = TRUE)

Arguments

obs
A times series (xts) object with a POSIXct index.

dt
New time interval, hours.

is.rate
If TRUE then these are rates i.e m/hr. Otherwise they are absolute values across the interval and are scaled before return by a factor equal to the ratio of the old interval to the new interval.

Value

An xts object with the new timestep

Details

Time series of observation data are often of different temporal resolutions, however the input to most hydrological models, as is the case with the Dynamic TOPMODEL, requires those data at the same interval. This provides a method to resample a collection of such data to a single interval.

Because of the methods used the results: - are not accurate when the input data does not have a constant timestep. The code issued a warnign and proceeds assuming the data are equally spaced with the modal timestep. - do not guarentee the requested time step but returns a series with the timestep computed from an integer rounding the ratio of the current and requested time step.

Examples

# Resample Brompton rainfall to 15 minute intervals
require(dynatop)
data("brompton")

obs <- resample_xts(brompton$rain, dt=15/60)

#> Warning: Irregularly spaced time series supplied to resample_xts - proceding with modal timestep, results are questionable

# check totals for Sept - Oct 2012
sum(obs$rain*15/60, na.rm=TRUE)

#> [1] 0.317

sum(brompton$rain, na.rm=TRUE)

#> [1] 0.317