Function to apply variable time delay routing

This function makes a model along with the specification of a point and inputs to the channel. It applies a time delay historgram routing based on the channel length and velocities.

time_delay_routing(model, channel_inflows, point_inflows = NULL,
  diffuse_inflows = NULL, initial_conditions = list())

Arguments

model
a dynamic TOPMODEL list object

channel_inflows
an xts object of flows to the channels

point_inflows
and xts object of point inflows

diffuse_inflows
an xts object of diffuse inflows

initial_conditions
a list of initial conditions

Value

An xts object of flows at the guage locations

Details

Timings are taken from the channel_inflows, other inputs are ignored unless there timings match