R6 Class for Dynamic TOPMODEL

R6 Class for Dynamic TOPMODEL

R6 Class for Dynamic TOPMODEL

Methods

Public methods

dynatop$new()

dynatop$add_data()

dynatop$clear_data()

dynatop$initialise()

dynatop$initialise_channel()

dynatop$sim_hillslope()

dynatop$sim_channel()

dynatop$sim()

dynatop$get_channel_inflow()

dynatop$plot_channel_inflow()

dynatop$get_gauge_flow()

dynatop$plot_gauge_flow()

dynatop$get_obs_data()

dynatop$get_model()

dynatop$get_mass_errors()

dynatop$get_states()

dynatop$plot_state()

dynatop$clone()

Method new()

Creates a dynatop class object from the a list based model description as generated by dynatopGIS.

Usage

dynatop$new(model, use_states = FALSE, delta = 1e-13)

Arguments

model

a dynamic TOPMODEL list object

use_states

logical if states should be imported

delta

error term in checking redistribution sums

drop_map

logical if the map should be dropped

Details

This function makes some basic consistency checks on a list representing a dynamic TOPMODEL model. The checks performed and basic 'sanity' checks. They do not check for the logic of the parameter values nor the consistency of states and parameters. Sums of the redistribution matrices are checked to be in the range 1 +/- delta.

Returns

invisible(self) suitable for chaining

Method add_data()

Adds observed data to a dynatop object

Usage

dynatop$add_data(obs_data)

Arguments

obs_data

an xts object of observed data

Details

This function makes some basic consistency checks on the observations to ensure they have uniform timestep and all required series are present.

Returns

invisible(self) suitable for chaining

Method clear_data()

Clears all forcing and simulation data except current states

Usage

dynatop$clear_data()

Returns

invisible(self) suitable for chaining

Method initialise()

Initialises a dynatop object in the most simple way possible.

Usage

dynatop$initialise(tol = 2 * .Machine$double.eps, max_it = 1000)

Arguments

tol

tolerance for the solution for the saturated zone

max_it

maximum number of iterations to use in the solution of the saturated zone

Returns

invisible(self) suitable for chaining

Method initialise_channel()

Initialises only the channel part of a dynatop object in the most simple way possible.

Usage

dynatop$initialise_channel()

Returns

invisible(self) suitable for chaining

Method sim_hillslope()

Simulate the hillslope output of a dynatop object

Usage

dynatop$sim_hillslope(
  keep_states = NULL,
  sub_step = NULL,
  tol = 2 * .Machine$double.eps,
  max_it = 1000,
  ftol = Inf
)

Arguments

keep_states

a vector of POSIXct objects (e.g. from xts) giving the time stamp at which the states should be kept

sub_step

simulation timestep in seconds, default value of NULL results in data time step

tol

tolerance on width of bounds in the solution for the saturated zone

max_it

maximum number of iterations to use in the solution of the saturated zone

ftol

tolerance in closeness to 0 in the solution for the saturated zone

Details

Both saving the states at every timestep and keeping the mass balance can generate very large data sets!! While ftol is implemented it is currently set to Inf to mimic the behaviour of previous versions. This will change in the future.

Returns

invisible(self) for chaining

Method sim_channel()

Simulate the channel output of a dynatop object

Usage

dynatop$sim_channel()

Returns

invisible(self) for chaining

Method sim()

Simulate the hillslope and channel components of a dynatop object

Usage

dynatop$sim(
  keep_states = NULL,
  sub_step = NULL,
  tol = 2 * .Machine$double.eps,
  max_it = 1000,
  ftol = Inf
)

Arguments

keep_states

a vector of POSIXct objects (e.g. from xts) giving the time stamp at which the states should be kept

sub_step

simulation timestep in seconds, default value of NULL results in data time step

tol

tolerance on width of bounds in the solution for the saturated zone

max_it

maximum number of iterations to use in the solution of the saturated zone

ftol

tolerance in closeness to 0 in the solution for the saturated zone

mass_check

Flag indicating is a record of mass balance errors should be kept

Details

Calls the sim_hillslope and sim_channel in sequence. Both saving the states at every timestep and keeping the mass balance can generate very large data sets!!

Returns

invisible(self) for chaining

Method get_channel_inflow()

Return channel inflow as an xts series or list of xts series

Usage

dynatop$get_channel_inflow(total = FALSE, separate = FALSE)

Arguments

total

logical if plot total inflow is to be plotted

separate

logical if the surface and saturated zone inflows should be returned separately

Method plot_channel_inflow()

Plot the channel inflow

Usage

dynatop$plot_channel_inflow(total = FALSE, separate = FALSE)

Arguments

total

logical if total inflow is to be plotted

separate

logical logical if the surface and saturated zone inflows should be plotted separately

Method get_gauge_flow()

Return flow at the gauges as an xts series

Usage

dynatop$get_gauge_flow(gauge = colnames(private$time_series$gauge_flow))

Arguments

gauge

names of gauges to return (default is all gauges)

Method plot_gauge_flow()

Get the flow at gauges

Usage

dynatop$plot_gauge_flow(gauge = colnames(private$time_series$gauge_flow))

Arguments

gauge

names of gauges to return (default is all gauges)

Method get_obs_data()

Get the observed data

Usage

dynatop$get_obs_data()

Method get_model()

Return the model

Usage

dynatop$get_model()

Method get_mass_errors()

Return the model

Usage

dynatop$get_mass_errors()

Method get_states()

Return states

Usage

dynatop$get_states(record = FALSE)

Arguments

record

logical TRUE if the record should be returned. Otherwise the current states returned

Method plot_state()

Plot a current state of the system

Usage

dynatop$plot_state(state, add_channel = TRUE)

Arguments

state

the name of the state to be plotted

add_channel

Logical indicating if the channel should be added to the plot

Method clone()

The objects of this class are cloneable with this method.

Usage

dynatop$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## the vignettes contains further details of the method calls.

data("Swindale") ## example data
ctch_mdl <- dynatop$new(Swindale$model) ## create with model
#> Warning: The following maps are missing:
#> hillslope, channel, channel_id
ctch_mdl$add_data(Swindale$obs) ## add observations
ctch_mdl$initialise() ## initialise model
ctch_mdl$sim() ## simulate model
#> Warning: Courant number for surface zone is over 0.7
#> Suggest maximum sub step is: 252.42seconds
#> Warning: Courant number for saturated zone is over 0.7
#> Suggest maximum sub step is: 18.14seconds